Add files via upload
Web-Scarping using Python.

Explanation.

Web Scraping is the process of programmatically extracting some data from a website. In Python, we can easy write program to scrape website by using the Beautiful Soup and Requests module. Requests module can be used to access the desired website and Beautiful Soup module can be used to parse the HTML source code of the website. Beautiful soup makes it very easy to parse HTML content and then provides multiple methods which can be used to extract the data from any HTML tags. All the website are written in HTML language hence in order to perform web scrapping, our program needs to read the HTML content. This is were Beautiful Soup comes into picture to make it very easy to access the contents of HTML using very simple methods.

In this small project, I wrote a  code to scrap the information about the movies from the movie website, which will be then loaded  to the .csv files(excel files) based on the different classes I have created, this can be done any website with this code reference by changing the parameters according to that particular website. This really helps  about the basic understanding how python code works on the html pages, parse the html and taking the specific information from the particular website.

U can download the code and run on any platform like pycharm, jupyter notebook and check the results.